amp: add burst start / end messages
Client needs information weather to updated UI immediately or to wait for more
messages. With these two more messages it is now clear that between burst start
and end messages it is not necessary to update UI. It is better to just collect
messages in that period and proceed with update after burst is ended.

Burst start message is always followed with the burst end in the future.