add fix for geoip2 file loading, passing of requests on load failure
Also:
add logic for doing a HEAD request to maxmind db source, to check
what is the value of Last-Modified header - to avoid unnecessary
downloads.

add option to define rules on when the database update should occur. Update is no longer done on a fixed time interval defined in geo package itself (GeoIPCheckInterval).